update logging-config configmap with empty values for Alloy (act as a placholder configmap)
